[ADDED] Update client config

+++ internal/pds-svc/model/client_config.go
@@ -9,7 +9,7 @@ type ClientConfig struct {
 	XID           string          `db:"xid"`
 	Key           string          `db:"key"`
 	Value         json.RawMessage `db:"value"`
-	ApplicationId int             `db:"applicationId"`
+	ApplicationId int64           `db:"applicationId"`
 	Metadata      json.RawMessage `db:"metadata"`
 	ItemMetadata
 }
